+++ config/splat.pspeu.stnp3.yaml
@@ -81,7 +81,7 @@ segments:
       - [0x1E6D8, c, np3/e_room_fg]
       - [0x1E7F0, c, np3/e_owl_knight]
       - [0x223D8, c, np3/st_common]
-      - [0x23E08, c, np3_psp/giantbro_helpers]
+      - [0x23E08, c, np3/giantbro_helpers]
       - [0x24870, c, np3_psp/e_blade]
       - [0x2A0B8, c, np3/e_zombie]
       - [0x2A588, c, np3/prim_helpers]
@@ -117,7 +117,7 @@ segments:
       - [0x58C30, .data, np3/e_room_fg]
       - [0x58CD0, .data, np3/e_owl_knight]
       - [0x58FC0, .data, np3/st_common]
-      - [0x591C0, .data, np3_psp/giantbro_helpers]
+      - [0x591C0, .data, np3/giantbro_helpers]
       - [0x59200, data]
       - [0x5A2F0, .data, np3/e_zombie]
       - [0x5A380, .rodata, np3/e_stage_name]
@@ -126,7 +126,7 @@ segments:
       - [0x5A410, .rodata, np3/e_collect]
       - [0x5A448, .rodata, np3/e_misc]
       - [0x5A470, .rodata, np3/water_effects]
-      - [0x5A490, .rodata, np3_psp/giantbro_helpers]
+      - [0x5A490, .rodata, np3/giantbro_helpers]
       - [0x5A4F0, .rodata, np3_psp/e_blade]
       - [0x5A580, bss, 5A580]
       - [0x5A7F0, .bss, np3/e_background_bushes_trees]
