+++ src/st/np3_psp/e_blade.c
